Feat/wrapper notification center

+++ personalization-sdk/src/main/kotlin/com/personalization/sdk/domain/usecases/notification/GetNotificationSourceUseCase.kt
@@ -8,7 +8,7 @@ class GetNotificationSourceUseCase @Inject constructor(
 ) {
 
     fun invoke(
-        timeDuration: Int
+        timeDuration: Long
     ) = notificationRepository.getNotificationSource(
         timeDuration = timeDuration
     )
